spring boot从2升级到3.4.5。删除未被使用的依赖。升级rocketmq-starter和mybatis-plus-starter，以及部分工具类包。

+++ peanut-chat-app-domain/src/main/java/com/sici/chat/model/chat/message/dto/MessageSendDTO.java
@@ -1,7 +1,5 @@
 package com.sici.chat.model.chat.message.dto;
 
-import com.alibaba.fastjson.annotation.JSONField;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -21,5 +19,5 @@ import java.io.Serializable;
 @AllArgsConstructor
 @NoArgsConstructor
 public class MessageSendDTO implements Serializable {
-    private Integer msgId;
+    private Long msgId;
 }
