+++ peanut-chat-app-server/src/main/java/com/sici/chat/builder/cache/RoomMemberCacheKeyBuilder.java
@@ -18,9 +18,9 @@ import static com.sici.common.constant.redis.key.RedisKeyConstant.ALL_KEY_SPLIT_
  */
 
 @Component
-public class RoomMemberCacheKeyBuilder implements RedisKeyBuilder<Integer, String> {
+public class RoomMemberCacheKeyBuilder implements RedisKeyBuilder<Long, String> {
     @Override
-    public String build(Integer req) {
+    public String build(Long req) {
         return ALL_KEY_PREFIX + ALL_KEY_SPLIT_ITEM + "room-friend"
                 + ALL_KEY_SPLIT_ITEM + req;
     }
