+++ peanut-chat-app-server/src/main/java/com/sici/chat/event/UserOfflineEvent.java
@@ -13,8 +13,8 @@ import org.springframework.context.ApplicationEvent;
 
 public class UserOfflineEvent extends ApplicationEvent {
     // TODO: 定义用户事件详细信息UserEventInfo(建表来保存用户上线和下线记录)  || created by 20148 at 12/4/2024 2:54 PM
-    private Integer userId;
-    public UserOfflineEvent(Object source, Integer userId) {
+    private Long userId;
+    public UserOfflineEvent(Object source, Long userId) {
         super(source);
         this.userId = userId;
     }
